requirements.txt
# Core LLM + embeddings
openai>=1.30.0
anthropic>=0.34.0

# RAG framework (used where it pulls weight; dropped where it doesn't)
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
langchain-mongodb>=0.1.0

# Vector stores
chromadb>=0.5.0           # local-first, used for step 01
pymongo>=4.7.0            # for MongoDB Atlas path (step 02+)
# pgvector alternative — install if using Postgres path
# psycopg[binary]>=3.1.0
# pgvector>=0.2.0

# Graph RAG (step 04)
neo4j>=5.20.0

# Keyword search (step 03)
rank-bm25>=0.2.2

# Eval + observability
langsmith>=0.1.0

# Utils
python-dotenv>=1.0.0
pypdf>=4.0.0
tiktoken>=0.7.0
pydantic>=2.7.0
